Move ComparisonMixin into conditions #66
