+++ tests/unit/test_conditions.py
@@ -34,6 +34,15 @@ from ..helpers.models import (
     conditions,
 )
 
+operations = [
+    (operator.ne, "!="),
+    (operator.eq, "=="),
+    (operator.lt, "<"),
+    (operator.le, "<="),
+    (operator.gt, ">"),
+    (operator.ge, ">=")
+]
+
 # Columns are sorted by model name
 empty_user_condition = (
     User.age.is_(None) &
@@ -513,4 +522,81 @@ def test_on_saved(engine):
         User.name.is_({"S": "foo"})
     )
 
+
 # END TRACKING SIGNALS ========================================================================== END TRACKING SIGNALS
+
+
+# COMPARISON MIXIN ================================================================================== COMPARISON MIXIN
+
+
+def test_column_equals_alias_exists():
+    """
+    == and != should map to attribute_not_exists and attribute_exists
+    when compared to None
+    """
+    column = Column(Integer)
+
+    condition = column.is_(None)
+    assert isinstance(condition, AttributeExists)
+    assert condition.column is column
+    assert condition.negate is True
+
+    condition = column.is_not(None)
+    assert isinstance(condition, AttributeExists)
+    assert condition.column is column
+    assert condition.negate is False
+
+
+@pytest.mark.parametrize("op_func, op_name", operations, ids=repr)
+def test_column_comparison(op_func, op_name):
+    column = Column(Integer)
+    value = object()
+
+    condition = op_func(column, value)
+    assert condition.comparator == op_name
+    assert condition.column is column
+    assert condition.value is value
+
+
+def test_column_between():
+    lower, upper = object(), object()
+    column = Column(Integer)
+    condition = column.between(lower, upper)
+
+    assert isinstance(condition, Between)
+    assert condition.column is column
+    assert condition.lower is lower
+    assert condition.upper is upper
+
+
+def test_column_in():
+    values = [object() for _ in range(3)]
+    column = Column(Integer)
+    condition = column.in_(values)
+
+    assert isinstance(condition, In)
+    assert condition.column is column
+    assert condition.values == values
+
+
+def test_column_begins_with():
+    value = object
+    column = Column(Integer)
+    condition = column.begins_with(value)
+
+    assert isinstance(condition, BeginsWith)
+    assert condition.column is column
+    assert condition.value == value
+
+
+def test_column_contains():
+    value = object
+    column = Column(Integer)
+    condition = column.contains(value)
+
+    assert isinstance(condition, Contains)
+    assert condition.column is column
+    assert condition.value == value
+
+
+# END COMPARISON MIXIN ========================================================================== END COMPARISON MIXIN
